+++ lib/facebook_web_sdk.dart
@@ -1,3 +1,4 @@
 library fb;
 
-export 'src/fb_base.dart';
+export 'src/fb.dart';
+export 'src/fb_base.dart' show FbInitOption, UiResponseData;