+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.6
+
+refactoring of the package 
+
++ Add uiShareDialog function
+
 ## 0.0.5
 
 + Dart 2 stable